+++ source/core/interface/decoder.cpp
@@ -131,7 +131,7 @@ void openhtj2k_decoder_impl::invoke(std::vector<int32_t *> &buf, std::vector<uin
     throw std::exception();
   }
 
-  auto tileSet = MAKE_UNIQUE<j2k_tile[]>(numTiles.x * numTiles.y);
+  auto tileSet = MAKE_UNIQUE<j2k_tile[]>(static_cast<size_t>(numTiles.x) * numTiles.y);
   for (uint16_t i = 0; i < static_cast<uint16_t>(numTiles.x * numTiles.y); ++i) {
     tileSet[i].dec_init(i, main_header, reduce_NL);
   }
